+++ backend/interface/database_api/src/connection.rs
@@ -1,352 +1,18 @@
-use database_types::connection_description::SshTunnelDescription;
-use diesel::RunQueryDsl;
-use log::warn;
-use russh::Channel;
-use russh::ChannelMsg;
-use russh::client::Msg;
-#[cfg(feature = "mysql")]
-use sea_query::MysqlQueryBuilder;
-#[cfg(feature = "postgres")]
-use sea_query::PostgresQueryBuilder;
-use sea_query::QueryStatementWriter;
-use sea_query::TableCreateStatement;
-use std::sync::Arc;
-
-use database_types::connection_description::ConnectionDescription;
+use crate::orm_connection::OrmConnection;
 use database_types::connection_description::DatabaseBackend;
+use database_types::connection_description::{ConnectionDescription, SshTunnelDescription};
+use diesel::Connection;
 #[cfg(feature = "mysql")]
 use diesel::MysqlConnection;
 #[cfg(feature = "postgres")]
 use diesel::PgConnection;
-use diesel::{Connection, MultiConnection};
 use log::error;
+use std::sync::Arc;
+
 use russh::client;
 
-use crate::api::models;
-use crate::json_field_conversion::JsonFieldConversion;
-
-#[derive(MultiConnection)]
-pub enum OrmConnection {
-    #[cfg(feature = "mysql")]
-    MySql(MysqlConnection),
-    #[cfg(feature = "postgres")]
-    PostgreSql(PgConnection),
-}
-
-impl OrmConnection {
-    pub fn to_string<QueryType: QueryStatementWriter>(&self, sql_query: QueryType) -> String {
-        return match self {
-            #[cfg(feature = "postgres")]
-            Self::PostgreSql(_) => sql_query.to_string(PostgresQueryBuilder),
-            #[cfg(feature = "mysql")]
-            Self::MySql(_) => sql_query.to_string(MysqlQueryBuilder),
-        };
-    }
-
-    pub fn to_string2(&self, sql_query: TableCreateStatement) -> String {
-        return match self {
-            #[cfg(feature = "postgres")]
-            Self::PostgreSql(_) => sql_query.to_string(PostgresQueryBuilder),
-            #[cfg(feature = "mysql")]
-            Self::MySql(_) => sql_query.to_string(MysqlQueryBuilder),
-        };
-    }
-
-    pub fn load_member<QueryType: QueryStatementWriter>(
-        &mut self,
-        sql_query: QueryType,
-    ) -> Option<Vec<models::Member>> {
-        let sql_query_string = self.to_string(sql_query);
-
-        let query_result = diesel::sql_query(&sql_query_string).load::<models::Member>(self);
-
-        return match query_result {
-            Ok(result) => Some(result),
-            Err(error) => {
-                error!(
-                    "Executing query '{}' failed due '{}'",
-                    sql_query_string, error
-                );
-                return None;
-            }
-        };
-    }
-
-    pub fn execute_sql(&mut self, sql_query: TableCreateStatement) -> Option<usize> {
-        let sql_query_string = self.to_string2(sql_query);
-
-        let query_result = diesel::sql_query(&sql_query_string).execute(self);
-
-        return match query_result {
-            Ok(result) => Some(result),
-            Err(error) => {
-                error!(
-                    "Executing query '{}' failed due '{}'",
-                    sql_query_string, error
-                );
-                return None;
-            }
-        };
-    }
-}
-
-pub struct SshConnection {
-    channel: Channel<Msg>,
-    // FIXME Type of shell required?
-    sql_login_command: String,
-    password: String,
-    backend: DatabaseBackend,
-}
-
-struct FieldGenerator {
-    field_name: String,
-    column_name: String,
-    column_index: usize,
-    to_json: Box<dyn Fn(&str) -> serde_json::Value>,
-}
-
-impl SshConnection {
-    pub fn to_string<QueryType: QueryStatementWriter>(&self, sql_query: QueryType) -> String {
-        return match self.backend {
-            #[cfg(feature = "postgres")]
-            DatabaseBackend::PostgreSql => sql_query.to_string(PostgresQueryBuilder),
-            #[cfg(feature = "mysql")]
-            DatabaseBackend::MySql => sql_query.to_string(MysqlQueryBuilder),
-        };
-    }
-
-    pub fn to_string2(&self, sql_query: TableCreateStatement) -> String {
-        return match self.backend {
-            #[cfg(feature = "postgres")]
-            DatabaseBackend::PostgreSql => sql_query.to_string(PostgresQueryBuilder),
-            #[cfg(feature = "mysql")]
-            DatabaseBackend::MySql => sql_query.to_string(MysqlQueryBuilder),
-        };
-    }
-
-    fn split_sql_cells(output: Vec<String>) -> Vec<Vec<String>> {
-        output
-            .iter()
-            .map(|line| line.split('\t').map(ToOwned::to_owned).collect())
-            .collect()
-    }
-
-    async fn read_channel_output(&mut self) -> Vec<String> {
-        let mut opt_exit_status = None;
-        let mut stdout_buffer = Vec::new();
-        let mut stderr_buffer = Vec::new();
-        loop {
-            let Some(message) = self.channel.wait().await else {
-                break;
-            };
-            match message {
-                ChannelMsg::Data { data } => stdout_buffer.extend_from_slice(&data),
-                ChannelMsg::ExtendedData { data, ext: _ } => stderr_buffer.extend_from_slice(&data),
-                ChannelMsg::ExitStatus { exit_status } => opt_exit_status = Some(exit_status),
-                _ => {}
-            }
-        }
-
-        let close_result = self.channel.close().await;
-        match close_result {
-            Ok(_) => {}
-            Err(error) => error!("Closing SSH channel failed due '{}'", error),
-        }
-
-        if opt_exit_status.is_none() {
-            warn!(
-                "Command '{}' finished without exit code",
-                self.sql_login_command
-            );
-        }
-        let exit_status = opt_exit_status.unwrap();
-        if exit_status != 0 {
-            error!(
-                "Command '{}' finished with exit code '{}'",
-                self.sql_login_command, exit_status
-            );
-        }
-
-        warn!(
-            "Command '{}' yielded stderr: '{}'",
-            self.sql_login_command,
-            String::from_utf8_lossy(&stderr_buffer)
-        );
-
-        return String::from_utf8_lossy(&stdout_buffer)
-            .lines()
-            .map(|line| line.to_owned())
-            .collect();
-    }
-
-    /* Structures representing data of databases are generated by Diesel. Due to lack of support for non-ASCII
-     * characters in Rust member names (really or is it just Diesel?) non-ASCII characters are replaced by underscores.
-     * In contrast command line tools return results containing column names including all special characters. To make
-     * the structure member match the column names returned by command line tools, the JSON keys representing column
-     * names (in the JSON object representing the command line output) have to be sanitized the same way like Diesel
-     * does.
-     * FIXME Is the sanitization function for Diesel member accessible directly?
-     */
-    fn get_field_name(column_name: &str) -> String {
-        // FIXME Verify in which cases it is true that non ASCII characters are replaced by underscore
-        column_name
-            .chars()
-            .map(|c| if c.is_ascii() { c } else { '_' })
-            .collect()
-    }
-
-    fn get_json_value_generators<SerializationType>(
-        column_names: &Vec<String>,
-    ) -> Vec<FieldGenerator>
-    where
-        SerializationType: JsonFieldConversion,
-    {
-        column_names
-            .iter()
-            .enumerate()
-            .map(|(index, column_name)| FieldGenerator {
-                field_name: Self::get_field_name(column_name),
-                /* 2026-06-28: Currently the actual column name member is unused but shows the potential difference to
-                 * the Rust member name generated by Diesel
-                 */
-                column_name: column_name.clone(),
-                column_index: index,
-                to_json: SerializationType::get_json_value_generator(
-                    Self::get_field_name(column_name).as_str(),
-                ),
-            })
-            .collect()
-    }
-
-    fn convert_to_json(
-        field_generators: Vec<FieldGenerator>,
-        rows: &[Vec<String>],
-    ) -> Vec<serde_json::Value> {
-        if rows.iter().any(|row| row.len() != field_generators.len()) {
-            error!("There are rows of different size than the row of field names");
-            return vec![];
-        }
-
-        let mut json_objects: Vec<serde_json::Value> = vec![];
-        for row in rows {
-            let mut object_properties = serde_json::Map::new();
-
-            for generator in &field_generators {
-                object_properties.insert(
-                    generator.field_name.clone(),
-                    (generator.to_json)(row[generator.column_index].as_str()),
-                );
-            }
-
-            json_objects.push(serde_json::json!(object_properties));
-        }
-
-        json_objects
-    }
-
-    // Return row major cells, i.e. a vector of rows, where each row is a vector of coloumns
-    async fn read_cells(&mut self, sql_query: String) -> Option<Vec<Vec<String>>> {
-        let sql_login_result = self
-            .channel
-            .exec(true, self.sql_login_command.as_bytes())
-            .await;
-        match sql_login_result {
-            Ok(_) => {
-                // FIXME Check all write results
-                let _password_write_result = self
-                    .channel
-                    .data(format!("{}\n", self.password).as_bytes())
-                    .await;
-                let _sql_query_write_result = self
-                    .channel
-                    .data(format!("{};\n", sql_query).as_bytes())
-                    .await;
-                let _eof_write_result = self.channel.eof().await;
-
-                let channel_output_lines = self.read_channel_output().await;
-                Some(Self::split_sql_cells(channel_output_lines))
-            }
-            Err(error) => {
-                error!("Could not log in to database due '{}'", error);
-                None
-            }
-        }
-    }
-
-    pub async fn load_member<QueryType: QueryStatementWriter>(
-        &mut self,
-        sql_query: QueryType,
-    ) -> Option<Vec<models::Member>> {
-        let sql_query_string = self.to_string(sql_query);
-        self.read_cells(sql_query_string).await.map(|cells| {
-            let cells_split = cells.split_first();
-
-            if cells_split.is_none() {
-                error!("Could not create objects since column names could not be determined");
-                return None;
-            }
-
-            let (column_names, rows) = cells_split.unwrap();
-
-            let field_generators =
-                SshConnection::get_json_value_generators::<models::Member>(column_names);
-
-            let json_result = SshConnection::convert_to_json(field_generators, rows);
-            Some(
-                json_result
-                    .into_iter()
-                    .map(|value| serde_json::from_value::<models::Member>(value))
-                    .filter(|result| match result {
-                        Ok(_member) => true,
-                        Err(error) => {
-                            error!("Could not interpret some row as member due '{}'", error);
-                            false
-                        }
-                    })
-                    .map(Result::unwrap)
-                    .collect::<Vec<models::Member>>(),
-            )
-        })?
-    }
-
-    pub async fn execute_sql(&mut self, sql_query: TableCreateStatement) -> Option<usize> {
-        let sql_query_string = self.to_string2(sql_query);
-        self.read_cells(sql_query_string)
-            .await
-            .map(|cells| cells.len())
-    }
-}
-
-pub enum DbConnection {
-    OrmBased(OrmConnection),
-    SshBased(SshConnection),
-}
-
-impl DbConnection {
-    pub fn to_string<QueryType: QueryStatementWriter>(&self, sql_query: QueryType) -> String {
-        return match self {
-            Self::OrmBased(connection) => connection.to_string(sql_query),
-            Self::SshBased(connection) => connection.to_string(sql_query),
-        };
-    }
-
-    pub async fn load_member<QueryType: QueryStatementWriter>(
-        &mut self,
-        sql_query: QueryType,
-    ) -> Option<Vec<models::Member>> {
-        return match self {
-            Self::OrmBased(connection) => connection.load_member(sql_query),
-            Self::SshBased(connection) => connection.load_member(sql_query).await,
-        };
-    }
-
-    pub async fn execute_sql(&mut self, sql_query: TableCreateStatement) -> Option<usize> {
-        return match self {
-            Self::OrmBased(connection) => connection.execute_sql(sql_query),
-            Self::SshBased(connection) => connection.execute_sql(sql_query).await,
-        };
-    }
-}
+use crate::db_connection::DbConnection;
+use crate::ssh_connection::SshConnection;
 
 struct SshClient {}
