+++ backend/interface/database_api/src/lib.rs
@@ -1,9 +1,12 @@
 pub mod api;
 
 mod connection;
+mod db_connection;
 mod frb_generated;
 mod json_field_conversion;
+mod orm_connection;
 mod schema;
+mod ssh_connection;
 
 #[cfg(test)]
 mod test_database_common;
