+++ backend/interface/database_api/src/orm_connection.rs
@@ -1,0 +1,78 @@
+#[cfg(feature = "mysql")]
+use diesel::MysqlConnection;
+#[cfg(feature = "postgres")]
+use diesel::PgConnection;
+use diesel::{MultiConnection, RunQueryDsl};
+#[cfg(feature = "mysql")]
+use sea_query::MysqlQueryBuilder;
+#[cfg(feature = "postgres")]
+use sea_query::PostgresQueryBuilder;
+use sea_query::{QueryStatementWriter, TableCreateStatement};
+
+use crate::api::models;
+use log::error;
+
+#[derive(MultiConnection)]
+pub enum OrmConnection {
+    #[cfg(feature = "mysql")]
+    MySql(MysqlConnection),
+    #[cfg(feature = "postgres")]
+    PostgreSql(PgConnection),
+}
+
+impl OrmConnection {
+    pub fn to_string<QueryType: QueryStatementWriter>(&self, sql_query: QueryType) -> String {
+        return match self {
+            #[cfg(feature = "postgres")]
+            Self::PostgreSql(_) => sql_query.to_string(PostgresQueryBuilder),
+            #[cfg(feature = "mysql")]
+            Self::MySql(_) => sql_query.to_string(MysqlQueryBuilder),
+        };
+    }
+
+    pub fn to_string2(&self, sql_query: TableCreateStatement) -> String {
+        return match self {
+            #[cfg(feature = "postgres")]
+            Self::PostgreSql(_) => sql_query.to_string(PostgresQueryBuilder),
+            #[cfg(feature = "mysql")]
+            Self::MySql(_) => sql_query.to_string(MysqlQueryBuilder),
+        };
+    }
+
+    pub fn load_member<QueryType: QueryStatementWriter>(
+        &mut self,
+        sql_query: QueryType,
+    ) -> Option<Vec<models::Member>> {
+        let sql_query_string = self.to_string(sql_query);
+
+        let query_result = diesel::sql_query(&sql_query_string).load::<models::Member>(self);
+
+        return match query_result {
+            Ok(result) => Some(result),
+            Err(error) => {
+                error!(
+                    "Executing query '{}' failed due '{}'",
+                    sql_query_string, error
+                );
+                return None;
+            }
+        };
+    }
+
+    pub fn execute_sql(&mut self, sql_query: TableCreateStatement) -> Option<usize> {
+        let sql_query_string = self.to_string2(sql_query);
+
+        let query_result = diesel::sql_query(&sql_query_string).execute(self);
+
+        return match query_result {
+            Ok(result) => Some(result),
+            Err(error) => {
+                error!(
+                    "Executing query '{}' failed due '{}'",
+                    sql_query_string, error
+                );
+                return None;
+            }
+        };
+    }
+}
